refactor: move navigation dropdown into molecules & change name
