+++ app/components/ui/atoms/NavigationDropdown.md
@@ -1,6 +1,6 @@
 ```
 const H1 = require('@pubsweet/ui').H1;
-;<DropdownNavLayout
+;<NavigationDropdown
   navList={[
     {label: 'Contact eLife', link: '/'},
     {label: 'Editorial staff', link: '/#editorial'},
@@ -9,5 +9,5 @@ const H1 = require('@pubsweet/ui').H1;
 >
   <H1>Contact eLife</H1>
   <p>You can use the "Check Status" link under Manuscript Tasks to find information about your article and how long different parts of the editorial and review process can take. You can also read about the system's requirements regarding browsers, emails, and PDFs if you are having difficulties. If you have any further questions about submitting, submitted, or accepted articles, please contact our editorial or production teams:</p>
-</DropdownNavLayout>
+</NavigationDropdown>
 ```
